+++ controller/settingscontroller.php
@@ -130,6 +130,7 @@ class SettingsController extends Controller {
             "macros" => $this->config->GetCustomizationMacros(),
             "tagsEnabled" => App::isEnabled("systemtags"),
             "reviewDisplay" => $this->config->GetCustomizationReviewDisplay(),
+            "theme" => $this->config->GetCustomizationTheme(),
             "templates" => $this->GetGlobalTemplates()
         ];
         return new TemplateResponse($this->appName, "settings", $data, "blank");
@@ -220,7 +221,8 @@ class SettingsController extends Controller {
                                     $forcesave,
                                     $help,
                                     $toolbarNoTabs,
-                                    $reviewDisplay
+                                    $reviewDisplay,
+                                    $theme
                                     ) {
 
         $this->config->SetDefaultFormats($defFormats);
@@ -237,6 +239,7 @@ class SettingsController extends Controller {
         $this->config->SetCustomizationHelp($help);
         $this->config->SetCustomizationToolbarNoTabs($toolbarNoTabs);
         $this->config->SetCustomizationReviewDisplay($reviewDisplay);
+        $this->config->SetCustomizationTheme($theme);
 
         return [
             ];