+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - macro launch setting
 - opening file location from viewer
 - Catalan translation
+- theme setting
 
 ## Changed
 - fix editor lang